Sl.repara {mgcv}	R Documentation

Applying re-parameterization from log-determinant of penalty matrix to model matrix.

Description

INTERNAL routine to apply re-parameterization from log-determinant of penalty matrix, ldetS to model matrix, X, blockwise.

Usage


Sl.repara(rp, X, inverse = FALSE, both.sides = TRUE)

Arguments

rp
reparametrization.

X
if X is a matrix it is assumed to be a model matrix whereas if X is a vector it is assumed to be a parameter vector.

inverse
if TRUE an inverse re-parametrization is performed.

both.sides
if inverse==TRUE and both.sides==FALSE then the re-parametrization only applied to rhs, as appropriate for a choleski factor. If both.sides==FALSE, X is a vector and inverse==FALSE then X is taken as a coefficient vector (so re-parametrization is inverse of that for the model matrix).

Value

A re-parametrized version of X.

Author(s)

Simon N. Wood <simon.wood@r-project.org>.

[Package mgcv version 1.8-24 Index]