dDeta {mgcv}	R Documentation

Obtaining derivative w.r.t. linear predictor

Description

INTERNAL function. Distribution families provide derivatives of the deviance and link w.r.t. mu = inv_link(eta). This routine converts these to the required derivatives of the deviance w.r.t. eta, the linear predictor.

Usage


dDeta(y, mu, wt, theta, fam, deriv = 0)

Arguments

y
vector of observations.

mu
if eta is the linear predictor, mu = inv_link(eta). In a traditional GAM mu=E(y).

wt
vector of weights.

theta
vector of family parameters that are not regression coefficients (e.g. scale parameters).

fam
the family object.

deriv
the order of derivative of the smoothing parameter score required.

Value

A list of derivatives.

Author(s)

Simon N. Wood <simon.wood@r-project.org>.

[Package mgcv version 1.8-24 Index]