gam.fit2 {mgcv}	R Documentation

P-IRLS GAM estimation with GCV & UBRE derivative calculation

Description

Estimation of GAM smoothing parameters is most stable if optimization of the UBRE or GCV score is outer to the penalized iteratively re-weighted least squares scheme used to estimate the model given smoothing parameters.

This routine estimates a GAM given log smoothing paramaters, and evaluates the first derivative of the GCV and UBRE scores of the model with respect to the log smoothing parameters. Calculation of exact derivatives is generally faster than approximating them by finite differencing, as well as generally improving the reliability of GCV/UBRE score minimization.

Not normally called directly, but rather a service routine for gam.

Usage

gam.fit2(x, y, sp, S=list(),rS=list(),off, H=NULL, 
         weights = rep(1, nobs), start = NULL, etastart = NULL, 
         mustart = NULL, offset = rep(0, nobs), family = gaussian(), 
         control = gam.control(), intercept = TRUE,deriv=TRUE,
         gamma=1,scale=1)

Arguments

x 	 The model matrix for the GAM.

y 	 The response variable.

sp 	 The log smoothing parameters.

S 	 A list of penalty matrices. Typically penalty matrices contain only a smallish square sub-matrix which is non-zero: this is what is actually stored. off[i] indicates which parameter is the first one penalized by S[[i]].

rS 	 List of square roots of penalty matrices, each having as few columns as possible, but as many rows as there are parameters.

off 	 off[i] indicates which parameter S[[i]][1,1] relates to.

H 	 The fixed penalty matrix for the model.

weights 	 prior weights for fitting.

start 	 optional starting parameter guesses.

etastart 	 optional starting values for the linear predictor.

mustart 	 optional starting values for the mean.

offset 	 the model offset

family 	 the family - actually this routine would never be called with gaussian()

control 	 control list as returned from glm.control

intercept 	 does the model have and intercept, TRUE or FALSE

deriv 	 Should derivatives of the GCV and UBRE scores be calculated? TRUE or FALSE

gamma 	 The weight given to each degree of freedom in the GCV and UBRE scores can be varied (usually increased) using this parameter.

scale 	 The scale parameter - needed for the UBRE score.

Details

This routine is basically glm.fit with some modifications to allow (i) for quadratic penalties on the log likelihood; (ii) derivatives of the model coefficients with respect to log smoothing parameters to be obtained (by updating alongside the P-IRLS iteration) and (iii) derivatives of the GAM GCV and UBRE scores to be evaluated at convergence.

In addition the routine applies step halving to any step that increases the penalized deviance substantially.

The most costly parts of the calculation are performed by calls to compiled C code (which in turn calls LAPACK routines) in place of the compiled code that would usually perform least squares estimation on the working model in the IRLS iteration.

Estimation of smoothing parameters by optimizing GCV scores obtained at convergence of the P-IRLS iteration was proposed by O'Sullivan et al. (1986), and is here termed `outer' iteration.

Author(s)

Simon N. Wood simon.wood@r-project.org

The routine has been modified from glm.fit in R 2.0.1, written by the R core (see glm.fit for further credits).

References

O 'Sullivan, Yandall & Raynor (1986) Automatic smoothing of regression functions in generalized linear models. J. Amer. Statist. Assoc. 81:96-103.

http://www.stats.gla.ac.uk/~simon/

See Also

gam.fit, gam, mgcv, magic

[Package mgcv version 1.2-3 Index]