gam.setup {mgcv}	R Documentation

Generalized Additive Model set up.

Description

This is an internal function of package mgcv. It is called by gam to obtain the design matrix and penalty matrices for a GAM set up using penalized regression splines. This is done by calling a mixture of R routines and compiled C code. For further information on usuage see code for gam.

Author(s)

Simon N. Wood simon.wood@r-project.org

References

Gu and Wahba (1991) Minimizing GCV/GML scores with multiple smoothing parameters via the Newton method. SIAM J. Sci. Statist. Comput. 12:383-398

Wood, S.N. (2000) Modelling and Smoothing Parameter Estimation with Multiple Quadratic Penalties. J.R.Statist.Soc.B 62(2):413-428

Wood, S.N. (2003) Thin plate regression splines. J.R.Statist.Soc.B 65(1):95-114

http://www.stats.gla.ac.uk/~simon/

See Also

gam, gamm, magic, mgcv

[Package mgcv version 1.2-3 Index]