gamm.setup {mgcv}	R Documentation

Generalized Additive Mixed Model set up.

Description

This is an internal function of package mgcv. It is called by gamm to set up a generalized additive mixed model in a form suitable for fitting by calls to glmmPQL from the MASS library or lme from the nlme library. The main task is the representation of the smooth terms as random effects.

Author(s)

Simon N. Wood simon.wood@r-project.org

References

Wood, S.N. (2004a) Stable and efficient multiple smoothing parameter estimation for generalized additive models. Journal of the American Statistical Association.

Wood, S.N. (2004b) Low rank scale invariant tensor product smooths for Generalized Additive Mixed Models. Technical Report of the Department of Statistics, University of Glasgow, UK.

http://www.stats.gla.ac.uk/~simon/

See Also

gamm

[Package mgcv version 1.2-3 Index]