mgcv.control {mgcv}	R Documentation

Setting mgcv defaults

Description

This is an internal function of package mgcv which allows control of the numerical options for fitting a generalized ridge regression problem using routine mgcv.

Usage

mgcv.control(conv.tol=1e-7,max.half=20,target.edf=NULL,min.edf=-1)

Arguments

conv.tol 	 The convergence tolerance.

max.half 	 successive step halvings are employed if the Newton method and then the steepest descent backup fail to improve the UBRE/GCV score. This is how many to use before giving up.

target.edf 	 If this is non-null it indicates that cautious optimization should be used, which opts for the local minimum closest to the target model edf if there are multiple local minima in the GCV/UBRE score.

min.edf 	 Lower bound on the model edf. Useful for avoiding numerical problems at high smoothing parameter values. Negative for none.

Author(s)

Simon N. Wood simon.wood@r-project.org

References

Gu and Wahba (1991) Minimizing GCV/GML scores with multiple smoothing parameters via the Newton method. SIAM J. Sci. Statist. Comput. 12:383-398

Wood, S.N. (2000) Modelling and Smoothing Parameter Estimation with Multiple Quadratic Penalties. J.R.Statist.Soc.B 62(2):413-428

http://www.maths.bath.ac.uk/~sw283/

See Also

mgcv

[Package mgcv version 1.5-5 Index]