new.name {mgcv}	R Documentation

Obtain a name for a new variable that is not already in use

Description

gamm works by transforming a GAMM into something that can be estimated by lme, but this involves creating new variables, the names of which should not clash with the names of other variables on which the model depends. This simple service routine checks a suggested name against a list of those in use, and if neccesary modifies it so that there is no clash.

Usage

 new.name(proposed,old.names)

Arguments

proposed
a suggested name

old.names
An array of names that must not be duplicated

Value

A name that is not in old.names.

Author(s)

Simon N. Wood simon.wood@r-project.org

References

http://www.maths.bris.ac.uk/~sw15190/

See Also

gamm

Examples

require(mgcv)
old <- c("a","tuba","is","tubby")
new.name("tubby",old)

[Package mgcv version 1.8-12 Index]