null.space.dimension {mgcv}	R Documentation

The basis of the space of un-penalized functions for a TPRS

Description

The thin plate spline penalties give zero penalty to some functions. The space of these functions is spanned by a set of polynomial terms. null.space.dimension finds the dimension of this space, M, given the number of covariates that the smoother is a function of, d, and the order of the smoothing penalty, m. If m does not satisfy 2m>d then the smallest possible dimension for the null space is found given d and the requirement that the smooth should be visually smooth.

Usage

null.space.dimension(d,m)

Arguments

d
is a positive integer - the number of variables of which the t.p.s. is a function.

m
a non-negative integer giving the order of the penalty functional, or signalling that the default order should be used.

Details

Thin plate splines are only visually smooth if the order of the wiggliness penalty, m, satisfies 2m > d+1. If 2m<d+1 then this routine finds the smallest m giving visual smoothness for the given d, otherwise the supplied m is used. The null space dimension is given by:

M=(m+d-1)!/(d!(m-1)!

which is the value returned.

Value

An integer (array), the null space dimension M.

Author(s)

Simon N. Wood simon.wood@r-project.org

References

Wood, S.N. (2003) Thin plate regression splines. J.R.Statist.Soc.B 65(1):95-114

http://www.maths.bris.ac.uk/~sw15190/

See Also

tprs

Examples

require(mgcv)
null.space.dimension(2,0)

[Package mgcv version 1.8-12 Index]