smooth.info {mgcv}	R Documentation

Generic function to provide extra information about smooth specification

Description

Takes a smooth specification object and adds extra basis specific information to it before smooth constructor called. Default method returns supplied object unmodified.

Usage

smooth.info(object)

Arguments

object
is a smooth specification object

Details

Sometimes it is necessary to know something about a smoother before it is constructed, beyond what is in the initial smooth specification object. For example, some smooth terms could be set up as tensor product smooths and it is useful for bam to take advantage of this when discrete covariate methods are used. However, bam needs to know whether a smoother falls into this category before it is constructed in order to discretize its covariates marginally instead of jointly. Rather than bam having a hard coded list of such smooth classes it is preferable for the smooth specification object to report this themselves. smooth.info method functions are the means for achieving this. When interpreting a gam formula the smooth.info function is applied to each smooth specification object as soon as it is produced (in interpret.gam0).

Value

A smooth specification object, which may be modified in some way.

Author(s)

Simon N. Wood simon.wood@r-project.org

References

Wood S.N. (2017) Generalized Additive Models: An Introduction with R (2nd edition). Chapman and Hall/CRC Press.

See Also

bam, smooth.construct, PredictMat

Examples

# See smooth.construct examples
spec <- s(a,bs="re")
class(spec)
spec$tensor.possible
spec <- smooth.info(spec)
spec$tensor.possible

[Package mgcv version 1.8-38 Index]