smoothCon {mgcv}	R Documentation

Prediction/Construction wrapper functions for GAM smooth terms

Description

Wrapper functions for construction of and prediction from smooth terms in a GAM. The purpose of the wrappers is to allow user-transparant re-parameterization of smooth terms, in order to allow identifiability constraints to be absorbed into the parameterization of each term, if required.

Usage

smoothCon(object,data,knots,absorb.cons=FALSE)
PredictMat(object,data)

Arguments

object 	 is a smooth specification object or a smooth object.

data 	 A data frame containing the values of the (named) covariates at which the smooth term is to be evaluated.

knots 	 An optional data frame supplying any knot locations to be supplied for basis construction.

absorb.cons 	 Set to TRUE in order to have identifiability constraints absorbed into the basis.

Details

These wrapper functions exist to allow smooths specified using smooth.construct and Predict.matrix method functions to be re-parameterized so that identifiability constraints are no longer required in fitting. This is done in a user transparent manner, but is typically of no importance in use of GAMs.

The parameterization used by gam can be controlled via gam.control.

Value

From smoothCon a smooth object returned by the appropriate smooth.construct method function. If constraints are to be absorbed then the object will have an attributes "qrc" and "nCons", the qr decomposition of the constraint matrix (returned by qr) and the number of constraints, respectively: these are used in the re-parameterization.
For predictMat a matrix which will map the parameters associated with the smooth to the vector of values of the smooth evaluated at the covariate values given in object.

Author(s)

Simon N. Wood simon.wood@r-project.org

References

http://www.stats.gla.ac.uk/~simon/

See Also

gam.control, smooth.construct, Predict.matrix

[Package mgcv version 1.2-3 Index]