totalPenaltySpace {mgcv}	R Documentation

Obtaining (orthogonal) basis for null space and range of the penalty matrix

Description

INTERNAL function to obtain (orthogonal) basis for the null space and range space of the penalty, and obtain actual null space dimension components are roughly rescaled to avoid any dominating.

Usage


totalPenaltySpace(S, H, off, p)

Arguments

S
a list of penalty matrices, in packed form.

H
the coefficient matrix of an user supplied fixed quadratic penalty on the parameters of the GAM.

off
a vector where the i-th element is the offset for the i-th matrix.

p
total number of parameters.

Value

A list of matrix square roots such that S[[i]]=B[[i]]%*%t(B[[i]]).

Author(s)

Simon N. Wood <simon.wood@r-project.org>.

[Package mgcv version 1.8-24 Index]