Predict.matrix.cr.smooth {mgcv}	R Documentation

Predict matrix method functions

Description

The various built in smooth classes for use with gam have associate Predict.matrix method functions to enable prediction from the fitted model.

Usage

## S3 method for class 'cr.smooth':
Predict.matrix(object, data)
## S3 method for class 'cs.smooth':
Predict.matrix(object, data)
## S3 method for class 'cyclic.smooth':
Predict.matrix(object, data)
## S3 method for class 'pspline.smooth':
Predict.matrix(object, data)
## S3 method for class 'tensor.smooth':
Predict.matrix(object, data)
## S3 method for class 'tprs.smooth':
Predict.matrix(object, data)
## S3 method for class 'ts.smooth':
Predict.matrix(object, data)

Arguments

object 	 a smooth object, usually generated by a smooth.construct method having processed a smooth specification object generated by an s or te term in a gam formula.

data 	 A data frame containing the values of the (named) covariates at which the smooth term is to be evaluated. Exact requirements are as for smooth.construct and smooth.construct2

.

Details

The Predict matrix function is not normally called directly, but is rather used internally by predict.gam etc. to predict from a fitted gam model. See Predict.matrix for more details, or the specific smooth.construct pages for details on a particular smooth class.

Value

A matrix mapping the coeffients for the smooth term to its values at the supplied data values.

Author(s)

Simon N. Wood simon.wood@r-project.org

References

Wood S.N. (2006) Generalized Additive Models: An Introduction with R. Chapman and Hall/CRC Press.

Examples

## see smooth.construct
 

[Package mgcv version 1.5-5 Index]