anova.gam {mgcv}	R Documentation

Hypothesis tests related to GAM fits

Description

Performs hypothesis tests relating to one or more fitted gam objects. For a single fitted gam object, Wald tests of the significance of each parametric and smooth term are performed. Otherwise the fitted models are compared using an analysis of deviance table. The tests are usually approximate, unless the models are un-penalized. Simulation evidence suggests that best p-value performance results from using ML estimated smoothing parameters.

Usage

## S3 method for class 'gam':
anova(object, ..., dispersion = NULL, test = NULL,
                    alpha = 0, freq = FALSE)
## S3 method for class 'anova.gam':
print(x, digits = max(3, getOption("digits") - 3),...)

Arguments

object,... 	 fitted model objects of class gam as produced by gam().

x 	 an anova.gam object produced by a single model call to anova.gam().

dispersion 	 a value for the dispersion parameter: not normally used.

test 	 what sort of test to perform for a multi-model call. One of "Chisq", "F" or "Cp".

alpha 	 adjustment to degrees of freedom per estimated smoothing parameter for a term when called with a single model object. See summary.gam for details.

freq 	 whether to use frequentist or Bayesian approximations for single smooth term p-values. See summary.gam for details.

digits 	 number of digits to use when printing output.

Details

If more than one fitted model is provided than anova.glm is used. If only one model is provided then the significance of each model term is assessed using Wald tests: see summary.gam for details of the actual computations. In the latter case print.anova.gam is used as the printing method. Note that the p-values for smooth terms are approximate only: simulation evidence suggests that they work best with REML or ML smoothness selection.

Value

In the multi-model case anova.gam produces output identical to anova.glm, which it in fact uses.
In the single model case an object of class anova.gam is produced, which is in fact an object returned from summary.gam.
print.anova.gam simply produces tabulated output.

WARNING

P-values for smooth terms are only approximate.

Author(s)

Simon N. Wood simon.wood@r-project.org with substantial improvements by Henric Nilsson.

See Also

gam, predict.gam, gam.check, summary.gam

Examples

library(mgcv)
set.seed(0)
dat <- gamSim(5,n=200,scale=2)

b<-gam(y ~ x0 + s(x1) + s(x2) + s(x3),data=dat)
anova(b)
b1<-gam(y ~ x0 + s(x1) + s(x2),data=dat)
anova(b,b1,test="F")

[Package mgcv version 1.5-5 Index]