bug.reports.mgcv {mgcv}	R Documentation

Reporting mgcv bugs.

Description

mgcv works largely because many people have reported bugs over the years. If you find something that looks like a bug, please report it, so that the package can be improved. mgcv does not have a large development budget, so it is a big help if bug reports follow the following guidelines.

The ideal report consists of an email to simon.wood@r-project.org with a subject line including mgcv somewhere, containing

The results of running sessionInfo in the R session where the problem occurs. This provides platform details, R and package version numbers, etc.

A brief description of the problem.

Short cut and paste-able code that produces the problem, including the code for loading/generating the data (using standard R functions like load, read.table etc).

Any required data files. If you send real data it will only be used for the purposes of de-bugging.

Of course if you have dug deeper and have an idea of what is causing the problem, that is also helpful to know, as is any suggested code fix. (Don't send a fixed package .tar.gz file, however - I can't use this).

Author(s)

Simon N. Wood simon.wood@r-project.org

[Package mgcv version 1.8-24 Index]