family.mgcv {mgcv}	R Documentation

Distribution families in mgcv

Description

As well as the standard families documented in family (see also glm) which can be used with functions gam, bam and gamm, mgcv also supplies some extra families, most of which are currently only usable with gam. These are described here.

Details

The following families are in the exponential family given the value of a single parameter. They are usable with all modelling functions.

Tweedie An exponential family distribution for which the variance of the response is given by the mean response to the power p. p is in (1,2) and must be supplied. See tw to estimate p.

negbin The negative binomial. See nb to estimate the theta parameter of the negative binomial.

The following families are for regression type models dependent on a single linear predictor, and with a log likelihood which is a sum of independent terms, each coprresponding to a single response observation. Usable only with gam, with smoothing parameter estimation by "REML" or "ML" (the latter does not integrate the unpenalized and parameteric effects out of the marginal likelihood optimized for the smoothing parameters).

ocat for ordered categorical data.

tw for Tweedie distributed data, when the power parameter relating the variance to the mean is to be estimated.

nb for negative binomial data when the theta parameter is to be estimated.

betar for proportions data on (0,1) when the binomial is not appropriate.

scat scaled t for heavy tailed data that would otherwise be modelled as Gaussian.

ziP for zero inflated Poisson data, when the zero inflation rate depends simply on the Poisson mean.

The following families implement more general model classes. Usable only with gam and only with REML smoothing parameter estimation.

cox.ph the Cox Proportional Hazards model for survival data.

gaulss a Gaussian location-scale model where the mean and the standard deviation are both modelled using smooth linear predictors.

ziplss a ‘two-stage’ zero inflated Poisson model, in which 'potential-presence' is modelled with one linear predictor, and Poisson mean abundance given potential presence is modelled with a second linear predictor.

mvn multivariate normal additive models.

Author(s)

Simon N. Wood (s.wood@r-project.org) & Natalya Pya

[Package mgcv version 1.8-7 Index]