fix.family.link {mgcv}	R Documentation

Modify families for use in GAM fitting

Description

Generalized Additive Model fitting by `outer' iteration, requires extra derivatives of the variance and link functions to be added to family objects. The functions add what is needed.

Usage

fix.family.link(fam)
fix.family.var(fam)

Arguments

fam 	 A family.

Details

Outer iteration GAM estimation requires derivatives of the GCV/UBRE score, which are obtained by differentiating the P-IRLS GAM fitting iteration w.r.t. the model smoothing parameters. The expressions for the derivatives require the second derivative of the link w.r.t. the mean, and the first derivative of the variance function w.r.t. the mean. These functions add functions evaluating these quantities to a family.

Value

A family object with extra component functions dvar and d2link.

Author(s)

Simon N. Wood simon.wood@r-project.org

[Package mgcv version 1.2-3 Index]