full.score {mgcv}	R Documentation

GCV/UBRE score for use within nlm

Description

Evaluates GCV/UBRE score for a GAM, given smoothing parameters. The routine calls gam.fit to fit the model, and is usually called by nlm to optimize the smoothing parameters.

This is basically a service routine for gam, and is not usually called directly by users. It is only used in this context for GAMs fitted by outer iteration (see gam.outer) when the the outer method is "nlm.fd" (see gam argument optimizer).

Usage


full.score(sp,G,family,control,gamma,...)

Arguments

sp
The logs of the smoothing parameters

G
a list returned by mgcv:::gam.setup

family
The family object for the GAM.

control
a list returned be gam.control

gamma
the degrees of freedom inflation factor (usually 1).

...
other arguments, typically for passing on to gam.fit.

Value

The value of the GCV/UBRE score, with attribute "full.gam.object" which is the full object returned by gam.fit.

Author(s)

Simon N. Wood simon.wood@r-project.org

[Package mgcv version 1.8-24 Index]