gam.fit {mgcv}	R Documentation

GAM P-IRLS estimation with GCV/UBRE smoothness estimation

Description

This is an internal function of package mgcv. It is a modification of the function glm.fit, designed to be called from gam when perfomance iteration is selected (not the default). The major modification is that rather than solving a weighted least squares problem at each IRLS step, a weighted, penalized least squares problem is solved at each IRLS step with smoothing parameters associated with each penalty chosen by GCV or UBRE, using routine magic. For further information on usage see code for gam. Some regularization of the IRLS weights is also permitted as a way of addressing identifiability related problems (see gam.control). Negative binomial parameter estimation is supported.

The basic idea of estimating smoothing parameters at each step of the P-IRLS is due to Gu (1992), and is termed ‘performance iteration’ or 'performance oriented iteration'.

Usage

 

gam.fit(G, start = NULL, etastart = NULL, 
        mustart = NULL, family = gaussian(), 
        control = gam.control(),gamma=1,
        fixedSteps=(control$maxit+1),...) 


Arguments

G
An object of the type returned by gam when fit=FALSE.

start
Initial values for the model coefficients.

etastart
Initial values for the linear predictor.

mustart
Initial values for the expected response.

family
The family object, specifying the distribution and link to use.

control
Control option list as returned by gam.control.

gamma
Parameter which can be increased to up the cost of each effective degree of freedom in the GCV or AIC/UBRE objective.

fixedSteps
How many steps to take: useful when only using this routine to get rough starting values for other methods.

...
Other arguments: ignored.

Value

A list of fit information.

Author(s)

Simon N. Wood simon.wood@r-project.org

References

Gu (1992) Cross-validating non-Gaussian data. J. Comput. Graph. Statist. 1:169-179

Gu and Wahba (1991) Minimizing GCV/GML scores with multiple smoothing parameters via the Newton method. SIAM J. Sci. Statist. Comput. 12:383-398

Wood, S.N. (2000) Modelling and Smoothing Parameter Estimation with Multiple Quadratic Penalties. J.R.Statist.Soc.B 62(2):413-428

Wood, S.N. (2004) Stable and efficient multiple smoothing parameter estimation for generalized additive models. J. Amer. Statist. Ass. 99:637-686

See Also

gam.fit3, gam, magic

[Package mgcv version 1.8-24 Index]