gam.outer {mgcv}	R Documentation

Minimize GCV or UBRE score of a GAM using 'outer' iteration

Description

Estimation of GAM smoothing parameters is most stable if optimization of the UBRE or GCV score is outer to the penalized iteratively re-weighted least squares scheme used to estimate the model given smoothing parameters.

This routine optimizes a GCV or UBRE score in this way. Basically the GCV or UBRE score is evaluated for each trial set of smoothing parameters by estimating the GAM for those smoothing parameters. The score is minimized w.r.t. the parameters numerically, using optim or nlm. Exact derivatives of the score can be used by fitting with gam.fit2, which improves efficiency and reliability relative to relying solely on finite difference derivatives.

Not normally called directly, but rather a service routine for gam.

Usage

gam.outer(lsp,fscale,family,control,method,gamma,G)

Arguments

lsp 	 The log smoothing parameters.

fscale 	 Typical scale of the GCV or UBRE score.

family 	 the model family.

control 	 control argument to pass to gam.fit if pure finite differencing is being used.

method 	 method list reurned from gam.method. This defines the optimization method to use.

gamma 	 The degree of freedom inflation factor for the GCV/UBRE score.

G 	 List produced by gam.setup, containing most of what's needed to actually fit GAM.

Details

Estimation of smoothing parameters by optimizing GCV scores obtained at convergence of the P-IRLS iteration was proposed by O'Sullivan et al. (1986), and is here termed `outer' iteration.

Author(s)

Simon N. Wood simon.wood@r-project.org

References

O 'Sullivan, Yandall & Raynor (1986) Automatic smoothing of regression functions in generalized linear models. J. Amer. Statist. Assoc. 81:96-103.

http://www.stats.gla.ac.uk/~simon/

See Also

gam.fit2, gam, mgcv, magic

[Package mgcv version 1.2-3 Index]