gam.outer {mgcv}	R Documentation

Minimize GCV or UBRE score of a GAM using ‘outer’ iteration

Description

Estimation of GAM smoothing parameters is most stable if optimization of the smoothness selection score (GCV, GACV, UBRE/AIC, REML, ML etc) is outer to the penalized iteratively re-weighted least squares scheme used to estimate the model given smoothing parameters.

This routine optimizes a smoothness selection score in this way. Basically the score is evaluated for each trial set of smoothing parameters by estimating the GAM for those smoothing parameters. The score is minimized w.r.t. the parameters numerically, using newton (default), bfgs, optim or nlm. Exact (first and second) derivatives of the score can be used by fitting with gam.fit3. This improves efficiency and reliability relative to relying on finite difference derivatives.

Not normally called directly, but rather a service routine for gam.

Usage

gam.outer(lsp,fscale,family,control,method,optimizer,
          criterion,scale,gamma,G,...)

Arguments

lsp 	 The log smoothing parameters.

fscale 	 Typical scale of the GCV or UBRE/AIC score.

family 	 the model family.

control 	 control argument to pass to gam.fit if pure finite differencing is being used.

method 	 method argument to gam defining the smoothness criterion to use (but depending on whether or not scale known).

optimizer 	 The argument to gam defining the numerical optimization method to use.

criterion 	 Which smoothness selction criterion to use. One of "UBRE", "GCV", "GACV", "REML" or "P-REML".

scale 	 Supplied scale parameter. Positive indicates known.

gamma 	 The degree of freedom inflation factor for the GCV/UBRE/AIC score.

G 	 List produced by mgcv:::gam.setup, containing most of what's needed to actually fit a GAM.

... 	 other arguments, typically for passing on to gam.fit3 (ultimately).

Details

See Wood (2008) for full details on `outer iteration'.

Author(s)

Simon N. Wood simon.wood@r-project.org

References

Wood, S.N. (2008) Fast stable direct fitting and smoothness selection for generalized additive models. J.R.Statist.Soc.B 70(3):495-518

http://www.maths.bath.ac.uk/~sw283/

See Also

gam.fit3, gam, mgcv, magic

[Package mgcv version 1.5-5 Index]