gam.setup {mgcv}	R Documentation

Generalized additive model set up

Description

This is an internal function of package mgcv. It is called by gam to obtain the design matrix and penalty matrices for a GAM set up using penalized regression splines. This is done by calling a mixture of R routines and compiled C code. For further information on usuage see code for gam.

Author(s)

Simon N. Wood simon.wood@r-project.org

References

Wood S.N. (2006) Generalized Additive Models: An Introduction with R. Chapman and Hall/CRC Press.

See Also

gam, gamm, magic, mgcv

[Package mgcv version 1.4-1 Index]