gamm.setup {mgcv}	R Documentation

Generalized additive mixed model set up

Description

This is an internal function of package mgcv. It is called by gamm to set up a generalized additive mixed model in a form suitable for fitting by calls to gammPQL or lme from the nlme library. The main task is the representation of the smooth terms as random effects. The routine first calls gam.setup, then checks that a model estimable by gamm has been specified in the model formula, and then transforms the problem to a form suitable for mixed model estimation.

Author(s)

Simon N. Wood simon.wood@r-project.org

References

Wood, S.N. (2004) Stable and efficient multiple smoothing parameter estimation for generalized additive models. Journal of the American Statistical Association. 99:673-686

Wood, S.N. (2006a) Low rank scale invariant tensor product smooths for generalized additive mixed models. Biometrics 62(4):1025-1036

Wood S.N. (2006b) Generalized Additive Models: An Introduction with R. Chapman and Hall/CRC Press.

http://www.maths.bath.ac.uk/~sw283/

See Also

gamm

[Package mgcv version 1.4-1 Index]