influence.gam {mgcv}	R Documentation

Extract the diagonal of the influence/hat matrix for a GAM

Description

Extracts the leading diagonal of the influence matrix (hat matrix) of a fitted gam object.

Usage


## S3 method for class 'gam'
influence(model,...)

Arguments

model
fitted model objects of class gam as produced by gam().

...
un-used in this case

Details

Simply extracts hat array from fitted model. (More may follow!)

Value

An array (see above).

Author(s)

Simon N. Wood simon.wood@r-project.org

See Also

gam

[Package mgcv version 1.8-24 Index]