interpret.gam {mgcv}	R Documentation

Interpret a GAM formula

Description

This is an internal function of package mgcv. It is a service routine for gam which splits off the strictly parametric part of the model formula, returning it as a formula, and interprets the smooth parts of the model formula.

Not normally called directly.

Usage

interpret.gam(gf)

Arguments

gf 	 A GAM formula as supplied to gam or gamm.

Value

An object of class split.gam.formula with the following items:

pf 	 A model formula for the strictly parametric part of the model.

pfok 	 TRUE if there is a pf formula.

smooth.spec 	 A list of class xx.smooth.spec objects where xx depends on the basis specified for the term. (These can be passed to smooth constructor method functions to actually set up penalties and bases.)

full.formula 	 An expanded version of the model formula in which the options are fully expanded, and the options do not depend on variables which might not be available later.

fake.formula 	 A formula suitable for use in evaluating a model frame.

response 	 Name of the response variable.

Author(s)

Simon N. Wood simon.wood@r-project.org

References

http://www.maths.bath.ac.uk/~sw283/

See Also

gam gamm

[Package mgcv version 1.5-5 Index]