magic.post.proc {mgcv}	R Documentation

Auxilliary information from magic fit

Description

Obtains Bayesian parameter covariance matrix, frequentist parameter estimator covariance matrix, estimated degrees of freedom for each parameter and leading diagonal of influence/hat matrix, for a penalized regression estimated by magic.

Usage


magic.post.proc(X,object,w=NULL)

Arguments

X
is the model matrix.

object
is the list returned by magic after fitting the model with model matrix X.

w
is the weight vector used in fitting, or the weight matrix used in fitting (i.e. supplied to magic, if one was.). If w is a vector then its elements are typically proportional to reciprocal variances (but could even be negative). If w is a matrix then t(w)%*%w should typically give the inverse of the covariance matrix of the response data supplied to magic.

Details

object contains rV (V, say), and scale (s, say) which can be used to obtain the require quantities as follows. The Bayesian covariance matrix of the parameters is VV's. The vector of estimated degrees of freedom for each parameter is the leading diagonal of VV'X'W'WX where W is either the weight matrix w or the matrix diag(w). The hat/influence matrix is given by WXVV'X'W' .

The frequentist parameter estimator covariance matrix is VV'X'W'WXVV's: it is sometimes useful for testing terms for equality to zero.

Value

A list with three items:

Vb
the Bayesian covariance matrix of the model parameters.

Ve
the frequentist covariance matrix for the parameter estimators.

hat
the leading diagonal of the hat (influence) matrix.

edf
the array giving the estimated degrees of freedom associated with each parameter.

Author(s)

Simon N. Wood simon.wood@r-project.org

See Also

magic

[Package mgcv version 1.8-24 Index]