print.gam {mgcv}	R Documentation

Generalized Additive Model default print statement

Description

This is the default print statement for a GAM object. If you need a list of everything that is part of a gam object see gam, or use names(). The family (including link), model formula, and estimated degrees of freedom for each model term (plus total) are printed, as well as the minimized GCV or UBRE score, depending on which was used.

Author(s)

Simon N. Wood simon@state.gla.ac.uk

References

Gu and Wahba (1991) Minimizing GCV/GML scores with multiple smoothing parameters via the Newton method. SIAM J. Sci. Statist. Comput. 12:383-398

Wood, S.N. (2000) Modelling and Smoothing Parameter Estimation with Multiple Quadratic Penalties. J.R.Statist.Soc.B 62(2):413-428

Wood, S.N. (2003) Thin plate regression splines. J.R.Statist.Soc.B 65(1):95-114

http://www.stats.gla.ac.uk/~simon/

See Also

gam,mgcv

[Package mgcv version 1.2-3 Index]