print.gam {mgcv}	R Documentation

Print a Generalized Additive Model object.

Description

The default print method for a gam object.

Usage

## S3 method for class 'gam':
print(x, ...)

Arguments

x, ... 	 fitted model objects of class gam as produced by gam().

Details

Prints out the family, model formula, effective degrees of freedom for each smooth term, and optimized value of the smoothness selection criterion used. See gamObject (or names(x)) for a listing of what the object contains. summary.gam provides more detail.

Author(s)

Simon N. Wood simon.wood@r-project.org

References

Wood, S.N. (2006) Generalized Additive Models: An Introduction with R. CRC/ Chapmand and Hall, Boca Raton, Florida.

http://www.maths.bath.ac.uk/~sw283/

See Also

gam, summary.gam

[Package mgcv version 1.5-5 Index]