print.gam {mgcv}	R Documentation

Generalized Additive Model default print statement

Description

This is the default print statement for a GAM object. If you need a list of everything that is part of a gam object see gam, or use names(). The family (including link), model formula, and estimated degrees of freedom for each model term (plus total) are printed, as well as the minimized GCV or UBRE/AIC score, depending on which was used.

Author(s)

Simon N. Wood simon.wood@r-project.org

References

Wood, S.N. (2006) Generalized Additive Models: An Introduction with R. CRC/ Chapmand and Hall, Boca Raton, Florida.

http://www.maths.bath.ac.uk/~sw283/

See Also

gam,mgcv

[Package mgcv version 1.4-1 Index]