residuals.gam {mgcv}	R Documentation

Generalized Additive Model residuals

Description

Returns residuals for a fitted gam model object. Pearson, deviance, working and response residuals are available.

Usage

## S3 method for class 'gam':
residuals(object, type = c("deviance", "pearson","scaled.pearson", 
                                         "working", "response"),...)

Arguments

object 	 a gam fitted model object.

type 	 the type of residuals wanted.

... 	 other arguments.

Details

Response residuals are the raw residuals (data minus fitted values). Scaled Pearson residuals are raw residuals divided by the standard deviation of the data according to the model mean variance relationship and estimated scale parameter. Pearson residuals are the same, but multiplied by the square root of the scale parameter (so they are independent of the scale parameter): ((y-m)/V(m)^0.5, where y is data m is model fitted value and V is model mean-variance relationship.). Both are provided since not all texts agree on the definition of Pearson residuals. Deviance residuals simply return the deviance residuals defined by the model family. Working residuals are the residuals returned from model fitting at convergence.

There is a special function for gam objects because of a bug in the calculation of Pearson residuals in some earlier versions of residual.glm.

Value

An array of residuals.

Author(s)

Simon N. Wood simon.wood@r-project.org

See Also

gam

[Package mgcv version 1.5-5 Index]