sdiag {mgcv}	R Documentation

Extract or modify diagonals of a matrix

Description

Extracts or modifies sub- or super- diagonals of a matrix.

Usage


sdiag(A,k=0)
sdiag(A,k=0) <- value

Arguments

A
a matrix

k
sub- (negative) or super- (positive) diagonal of a matrix. 0 is the leading diagonal.

value
single value, or vector of the same length as the diagonal.

Value

A vector containing the requested diagonal, or a matrix with the requested diagonal replaced by value.

Author(s)

Simon N. Wood simon.wood@r-project.org

Examples


require(mgcv)
A <- matrix(1:35,7,5)
A
sdiag(A,1) ## first super diagonal
sdiag(A,-1) ## first sub diagonal

sdiag(A) <- 1 ## leading diagonal set to 1
sdiag(A,3) <- c(-1,-2) ## set 3rd super diagonal 


[Package mgcv version 1.8-24 Index]