step.gam {mgcv}	R Documentation

Alternatives to step.gam

Description

There is no step.gam in package mgcv. The mgcv default for model selection is to use MSE/KL-distance criteria such as GCV or UBRE/AIC. Since the smoothness estimation part of model selection is done in this way it is logically most consistent to perform model selection on the basis of such criteria: i.e. to decide which terms to include or omit by looking at changes in GCV/UBRE/AIC score.

To facilitate fully automatic model selection the package includes 2 classes of smoothers ("cs" and "ts": see s) which can be penalized to zero for sufficiently high smoothing parameter estimates: use of such smooths provides an effective alternative to step-wise model selection. The example below shows an example of the application of this approach, where selection is a fully integrated part of model estimation.

Author(s)

Simon N. Wood simon.wood@r-project.org

Examples

## an example of GCV based model selection as
## an alternative to stepwise selection
library(mgcv)
set.seed(0);n <- 400
dat <- gamSim(1,n=n,scale=2)
dat$x4 <- runif(n, 0, 1)
dat$x5 <- runif(n, 0, 1)
attach(dat)
## Note the increased gamma parameter below to favour
## slightly smoother models...
b<-gam(y~s(x0,bs="ts")+s(x1,bs="ts")+s(x2,bs="ts")+
   s(x3,bs="ts")+s(x4,bs="ts")+s(x5,bs="ts"),gamma=1.4)
summary(b)
plot(b,pages=1)
detach(dat);rm(dat)

[Package mgcv version 1.4-1 Index]